update decoding interfaces

+++ src/stavisky_lab_to_nwb/braintotext/braintotext_convert_session.py
@@ -18,7 +18,7 @@ def session_to_nwb(port: int, host: str, output_dir_path: Union[str, Path], stub
 
     # Extract session metadata
     rdb_metadata = r.xrange("metadata")[0][1]
-    session_start_time = np.frombuffer(rdb_metadata[b"startTime"], dtype=np.float64).item()
+    session_start_time = np.frombuffer(rdb_metadata[b"startTime"], dtype=np.int64).astype(float).item()
 
     # Prepare output path
     output_dir_path = Path(output_dir_path)
